UML-3460 renamed datastore to lpa store
