+++ service-api/app/src/App/src/Entity/Casters/ExtractCountryFromLpaStore.php
@@ -9,7 +9,7 @@ use EventSauce\ObjectHydrator\ObjectMapper;
 use EventSauce\ObjectHydrator\PropertyCaster;
 
 #[Attribute(Attribute::TARGET_PARAMETER)]
-class ExtractCountryFromDataStore implements PropertyCaster
+class ExtractCountryFromLpaStore implements PropertyCaster
 {
     public function cast(mixed $value, ObjectMapper $hydrator): ?string
     {
