--- service-api/app/src/App/src/Entity/DataStore/DataStoreAttorney.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Entity\DataStore;
-
-class DataStoreAttorney extends DataStorePerson
-{
-}
