+++ service-api/app/src/App/src/Entity/LpaStore/LpaStoreAttorney.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Entity\LpaStore;
+
+class LpaStoreAttorney extends LpaStorePerson
+{
+}
