+++ service-api/app/src/App/src/Entity/LpaStore/LpaStoreTrustCorporations.php
@@ -2,41 +2,38 @@
 
 declare(strict_types=1);
 
-namespace App\Entity\DataStore;
+namespace App\Entity\LpaStore;
 
-use App\Entity\Casters\ExtractAddressLine1FromDataStore;
-use App\Entity\Casters\ExtractCountryFromDataStore;
-use App\Entity\Casters\ExtractTownFromDataStore;
-use App\Entity\Person;
-use DateTimeImmutable;
+use App\Entity\Casters\ExtractAddressLine1FromLpaStore;
+use App\Entity\Casters\ExtractCountryFromLpaStore;
+use App\Entity\Casters\ExtractTownFromLpaStore;
+use DateTimeInterface;
 use EventSauce\ObjectHydrator\MapFrom;
 
-class DataStorePerson extends Person
+class LpaStoreTrustCorporations extends LpaStorePerson
 {
     public function __construct(
+        #[MapFrom('name')]
+        public readonly ?string $companyName,
         ?string $name,
         #[MapFrom('address')]
-        #[ExtractAddressLine1FromDataStore]
+        #[ExtractAddressLine1FromLpaStore]
         ?string $addressLine1,
         ?string $addressLine2,
         ?string $addressLine3,
         #[MapFrom('address')]
-        #[ExtractCountryFromDataStore]
+        #[ExtractCountryFromLpaStore]
         ?string $country,
         ?string $county,
         ?string $postcode,
         #[MapFrom('address')]
-        #[ExtractTownFromDataStore]
+        #[ExtractTownFromLpaStore]
         ?string $town,
         ?string $type,
-        #[MapFrom('dateOfBirth')]
-        ?DateTimeImmutable $dob,
+        ?DateTimeInterface $dob,
         ?string $email,
-        #[MapFrom('firstname')]
         ?string $firstname,
-        #[MapFrom('firstNames')]
         ?string $firstnames,
-        #[MapFrom('lastName')]
         ?string $surname,
         ?string $otherNames,
         #[MapFrom('status')]
@@ -58,7 +55,7 @@ class DataStorePerson extends Person
             $firstnames,
             $surname,
             $otherNames,
-            $systemStatus,
+            $systemStatus
         );
     }
 }
